refactor: extract latLonToVector3 as utility function
