+++ src/utils/geoUtils.ts
@@ -1,4 +1,5 @@
 import { Feature, GeoJSONGeometry, GeoDataType } from "../types/geo.ts";
+import * as THREE from "three";
 
 export function createGeometryArray(json: GeoDataType): GeoJSONGeometry[] {
   const geometry_array: GeoJSONGeometry[] = [];
@@ -65,4 +66,19 @@ export function createCoordinateArray(feature: number[][]): number[][] {
   }
 
   return result;
+}
+
+// 위도/경도를 3D 좌표 (Vector3)로 변환하는 함수
+export function latLonToVector3(
+  lat: number,
+  lon: number,
+  radius: number
+): THREE.Vector3 {
+  const phi = (90 - lat) * (Math.PI / 180);
+  const theta = (lon + 180) * (Math.PI / 180);
+  return new THREE.Vector3(
+    -radius * Math.sin(phi) * Math.cos(theta),
+    radius * Math.cos(phi),
+    radius * Math.sin(phi) * Math.sin(theta)
+  );
 }
